Deleted key value from hash
Deleted all key values from the hash and fixed all the errors in result
of that deletion. Also had to move the Handle class to be its own class
instead of a subclass in order for it to work with the Hash class. Also
created a new method within the MemoryManager to handle the conversion
from handle to string. The test cases for the hash table has to be fixed
though.